+++ crates/amplifier-api/src/lib.rs
@@ -5,8 +5,7 @@ mod client;
 pub use client::*;
 mod error;
 pub use error::AmplifierApiError;
+pub mod big_int;
 pub mod types;
 pub mod util;
-/// Big integer type with configurable underlying type via features
-pub mod bigint;
 pub use chrono;
