+++ crates/amplifier-api/src/types.rs
@@ -3,7 +3,6 @@
 
 use core::fmt::{Display, Formatter};
 
-pub use big_int::BigInt;
 use borsh::{BorshDeserialize, BorshSerialize};
 use chrono::{DateTime, Utc};
 pub use id::*;
@@ -11,6 +10,8 @@ use infrastructure::interfaces::publisher::QueueMsgId;
 use serde::{Deserialize, Serialize};
 use typed_builder::TypedBuilder;
 pub use {bnum, uuid};
+
+use crate::big_int::BigInt;
 
 /// Represents an address as a non-empty string.
 pub type Address = String;
@@ -389,8 +390,8 @@ pub struct Token {
     pub token_id: Option<TokenId>,
     /// the amount in token’s denominator
     #[borsh(
-        serialize_with = "crate::util::serialize_bigint",
-        deserialize_with = "crate::util::deserialize_bigint"
+        serialize_with = "crate::big_int::serialize",
+        deserialize_with = "crate::big_int::deserialize"
     )]
     pub amount: BigInt,
 }
@@ -1185,47 +1186,6 @@ pub struct ErrorResponse {
     pub request_id: Option<RequestId>,
 }
 
-mod big_int {
-    use serde::{Deserialize, Deserializer, Serialize};
-
-    /// Represents a big integer as a string matching the pattern `^(0|[1-9]\d*)$`.
-    #[derive(Clone, Debug, PartialEq, Eq)]
-    pub struct BigInt(pub bnum::types::I512);
-    impl BigInt {
-        /// Creates a new [`BigInt`].
-        #[must_use]
-        pub const fn new(num: bnum::types::I512) -> Self {
-            Self(num)
-        }
-
-        /// Helper utility to transform u64 into a `BigInt`
-        #[must_use]
-        pub fn from_u64(num: u64) -> Self {
-            Self(num.into())
-        }
-    }
-
-    impl Serialize for BigInt {
-        fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-        where
-            S: serde::Serializer,
-        {
-            let string = self.0.to_string();
-            serializer.serialize_str(&string)
-        }
-    }
-
-    impl<'de> Deserialize<'de> for BigInt {
-        fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-        where
-            D: Deserializer<'de>,
-        {
-            let string = String::deserialize(deserializer)?;
-            let number = bnum::types::I512::parse_str_radix(string.as_str(), 10);
-            Ok(Self(number))
-        }
-    }
-}
 /// reference types are copied from the following documentation [link](https://bright-ambert-2bd.notion.site/Amplifier-GMP-API-EXTERNAL-911e740b570b4017826c854338b906c8#e8a7398607bd496eb0b8e95e887d6574)
 #[cfg(test)]
 mod tests {
@@ -1235,6 +1195,7 @@ mod tests {
     use simd_json::{from_slice, json, to_owned_value, to_string};
 
     use super::*;
+    use crate::big_int::BigInt;
 
     const BASE64_PAYLOAD: &str = "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAgAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAASaGVsbG8gdGVzdC1zZXBvbGlhAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA==";
     const BASE64_PAYLOAD_HASH: &str = "Y2YO3UuCRRackxbPYX9dWmNTYcnAMOommp9g4ydb3i4=";
